Implement DancerNotFoundException in DancerService. Closes #25.

+++ src/main/java/com/example/my_first_jpa_app/service/DancerServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.my_first_jpa_app.service;
 
+import com.example.my_first_jpa_app.controller.DancerNotFoundException;
 import com.example.my_first_jpa_app.domain.Dancer;
 import com.example.my_first_jpa_app.repository.DancerRepository;
 import lombok.AllArgsConstructor;
@@ -19,6 +20,9 @@ public class DancerServiceImpl implements DancerService {
 
     @Override
     public Dancer getDancer(Integer id) {
+        if(!dancerRepository.existsById(id)){
+            throw new DancerNotFoundException("Dancer with id: " + id + " was not found!");
+        }
         return dancerRepository.getReferenceById(id);
     }
 
@@ -30,7 +34,9 @@ public class DancerServiceImpl implements DancerService {
 
     @Override
     public Dancer updateDancer(Integer id, Dancer dancer) {
-
+        if(!dancerRepository.existsById(id)){
+            throw new DancerNotFoundException("Dancer with id: " + id + " was not found!");
+        }
         Dancer toUpdate = dancerRepository.getReferenceById(id);
         toUpdate.setName(dancer.getName());
         toUpdate.setDanceStyle(dancer.getDanceStyle());
@@ -39,6 +45,9 @@ public class DancerServiceImpl implements DancerService {
 
     @Override
     public Dancer deleteDancer(Integer id) {
+        if(!dancerRepository.existsById(id)){
+            throw new DancerNotFoundException("Dancer with id: " + id + " was not found!");
+        }
         Dancer deleted = dancerRepository.getReferenceById(id);
         dancerRepository.deleteById(id);
         return deleted;
